app: fix wrong format restriction for indexed images.
When setting a new palette, we must not use the new palette format
restriction (especially if it has none!), but the indexed image's
format, which is unchanged anyway. So move the call to
gimp_palette_restrict_format() into gimp_image_colormap_update_formats()
and use the correct format to restrict to.

See: https://gitlab.gnome.org/GNOME/gimp/-/issues/12108#note_2238182