app:vectors: fix #12120 CRITICAL on shift-start action on a bezier Handle
This eliminates the CRITICAL, which was harmless anyway.

As discussed in the issue, there are more issues:

1. The "shift-start" action is not documented in the user manual.

2. The "shift-start" action quietly does nothing in the case
it has nothing to do.  Better GUI would be to disable the menu item in that case.

3.  The "shift-start" action is non-intuitive, even bizarre.
It could probably be eliminated.
I myself don't understand the use case, and confused it for Path>Reverse.

An alternative fix is to disable the menu item when a handle is the operand.
This fix is better because it gives the user more places to RMB a context menu,
and the context menu mutates less.

Probably needs backport to 2.10.