app: show clearer we are working in sRGB space.
This was already the case because gegl_operation_get_format() had no
space (i.e. default, sRGB). it's not the source format.
But let's make it clearer because otherwise on a short skimming, it
looks like we are working on the source space.

Also adding a comment on why working on the source space is not a good
idea right now anyway (it's *very* slow for anything but sRGB).