core, widgets: Improve filter performance
After 59b7b6a5, we had reports that filters
ran much slower. This patch is an initial
attempt to improve performance.
It reduces calls to gegl_node_get_bounding_box (),
which may rerender the graph each time
it's called. We should now only try to
recalculate the crop when commiting or
resizing the filter.