authors: update.
I'm surprised that Idriss and Shubham were not already in the file!

Also adding Denis for his design work, and Sevenix for the splash
screen.

Bruno was added for the awesome work in packaging and CI. Even though
it's not GIMP itself, it definitely counts as code and original work.

I didn't add the new GSoC contributors because Varun's work is
essentially in GEGL (which has its own credits) and there was no major
original feature pushed by Cheesequake unfortunately, only bugfixes. So
I'm unsure if we can add them.