build/windows: Prepare Installer scripts to RCs versioning and beyond
Adds GIMP_MUTEX_VERSION var to make possible to change the way we handle
multiple GIMP installations without needing to manually change parts of
the script to MAJOR or GIMP_APP_VERSION like a zoombie.

Also, this reorganizes some sections of the .iss file (for the third time)
and make LZMA multi threading the default even for stable builds.