app: migrate also from ~/.var/ if the 2.10 folder is not found in XDG config home.
So this one is a complicated beast, because inside the flatpak's
sandbox, ~/.var/ will not be the same as on the host (even with home
permission). In particular, we won't be able to find the 2.10 config folder in
there when running GIMP 3.

Yet local ~/.var/app/org.gimp.GIMP/config/ is in fact be mounted to
/var/config/! So inside the sandbox, this is where we must look in to
find a previous 2.10 config folder.

With this change, both flatpak and non-flatpak GIMP should be able to
find a 2.10 config folder both in the host XDG_CONFIG_HOME and inside
~/.var/ by:

* First looking inside XDG_CONFIG_HOME;
* Then looking in ~/.var/app/org.gimp.GIMP/config/ (which in non-flatpak
  environment will find a former flatpak config folder);
* Finally looking in /var/config/, but only is /.flatpak-info exists (so
  we are in a flatpak, and in there, this is the mapping for the host
  ~/.var).

This is not tested inside the flatpak sandbox yet.