build/windows: Bundle and distribute gspawn*.exe to be able to open links
This file has stopped of being bundled in c808d13b as a consequence of a
bad practice of the past: not documenting the purpose of bundled files.

Let's bundle it again commenting the reason so the future generations can
understand why it is bundled: gspawn*-console is needed to open hyperlinks.
Of course, it does more than this, but one use case is enough to understand.

The warning "Couldn't export handle, unsupported windowing system" still will
be output on terminal since the fix (GNOME/gtk!7694) wasn't released yet.

---

Also, clarify commenting other bundled files and dirs in the bundling script.