pdb: Set choice args to default in gimp-file-save
In gimp_file_save (), we clear out the values
of any String parameters to prevent unexpected
results. However, GimpChoice is also a
String type. This causes the function to
fail because empty string is not a valid
option for GimpChoice.

This patch adds a prior check if the argument
is a GimpChoice, and sets it to its default
value instead to prevent this issue.