Fix #12360 SF server not returning same stream as version 2
Change SF server to load /scripts, just like extensio-script-fu.

This restores SF server to v2 behaviour re:
1) the content of responses
2) the need for a run_mode arg in calls to scripts

Thus calls to scripts (submitted to the server)
are evaluated without a call via the PDB to extension-script-fu.
This means such a call yields the string repr
of the last evaluated expression of the script,
instead of the string repr of a call via the PDB,
which is #t since scripts are declared in the PDB as returning void.

This also means the run-mode argument is not required in such calls,
since it is a call to a run_func in Scheme, which has no run_mode arg.

This also means that the SF server will continue to run
even if extension-script-fu crashes.

A test is that a call to a simple "echo" script, installed to /scripts,
yields the same response from the server in both GIMP 2 and 3.