ScriptFu: fix #10741 tooltip and PDB description have underscore
Elide the underscore char from the blurb of the GParamSpec.
The tooltip remains redundant with the menu label.
We can't use an empty string (omit the tooltip)
because the blurb is also the description in the PDB Browser,
and the generated name of the arg is not descriptive.

FUTURE: let script author also declare a tooltip,
and a non-generated but unique name of the arg.