app/xcf: Don't override channel opacity on XCF load
When loading channels, the alpha value is stored separately from
the RGB value as "opacity". After the switch to GeglColor, we were
overriding the opacity values by always using "R'G'B" format, so
the loaded opacity was always reset to 100%.

This patch pulls the opacity value before setting the color, then
restores the opacity afterwards.