ScriptFu: fix #12504, unhandled GUI events on MacOS
Previously, just hanging a read, extension-script-fu missed GUI events
that closed dialogs on MacOS.

The fix is a proper event loop for a plugin that persists, reads, and has GUI:
calling gimp_plug_in_persistent_enable and a main_loop,
instead of gimp_plug_in_persistent_process.

The fix is not conditional on MacOS, but seems events missed only on MacOS.

The previous fix for MacOS, #6577, only works for old-style GUI (script-fu-interface.c).
That was not the best fix, and doesn't work for new-style GUI (GimpProcedureDialog).
If symptoms still appear in 2.10.38
(#12150 says dialogs dissappear but a dock icon has spinning ball)
this should be backported to 2.10.